Reflecting recent commits to individual files.

+++ build.bat
@@ -1,4 +1,7 @@
+if [ -e "codebase.pde" ]
+then
+	rm codebase.pde
+fi
 cat *.pde > codebase.pde
 cp codebase.pde mario/
 cp codebase.pde docs/
-rm codebase.pde
